Implement support for RedisClient::Cluster#with
This commit introduces `Redis::Cluster#with`, which acts as a proxy for
`RedisClient::Cluster#with` and supports pinning a cluster client to a
particular connection. This allows WATCH/MULTI/EXEC transactions to then
be performed on the yielded connection without ambiguity.

The yielded connection object is a lightweight proxy that also
implements all of the Redis command DSL. This means that you can call
`#watch`, `#pipelined`, `#multi`, and others on the yielded proxy just
like you could with an ordinary `Redis` instance.